requirements.txt
pyspark
apache-airflow
apache-airflow-providers-amazon
boto3
pandas
hadoop-aws
python-dotenv:wq

